pyproject.toml
[tool.poetry]
name = "gitlab-search-bot"
version = "0.1.0"
description = ""
authors = ["Dylan Castillo <dylanjcastillo@gmail.com>"]
readme = "README.md"
packages = [{include = "gitlab_search_bot"}]

[tool.poetry.dependencies]
python = "^3.10,<3.11"
qdrant-client = "^1.0.4"
sentence-transformers = "^2.2.2"
cohere = "^3.10.0"
gunicorn = "^20.1.0"
fastapi = {extras = ["all"], version = "^0.94.0"}
python-dotenv = "^1.0.0"
mistletoe = "^1.0.1"
beautifulsoup4 = "^4.11.2"
bleach = "^6.0.0"
tiktoken = "^0.3.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
mypy = "^1.1.1"
notebook = "^6.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
